requirements.txt
psutil
requests
python-telegram-bot